move signature_v2.go to pkg/s3
Have the code in a central place to make it accessible to additional
callers.

Recommended git option to view this commit: --color-moved.
